Refactor color_switch() to not inject callbacks
In colorscheme mode, I was injecting callbacks that set the colorscheme.
This led to a few issues where loading a config failed. This change
refactors the color switch function to change the colorscheme without
injected callbacks.

Once that was done, it made sense to also only call the callbacks in
custom mode, so I refactored out the callbacks into a runner function.